+++ app/src/main/java/com/tpc/nudj/ui/screen/auth/splash/SplashScreen.kt
@@ -9,6 +9,7 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -17,40 +18,42 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.tpc.nudj.R
+import com.tpc.nudj.ui.components.NudjTopAppBar
 import com.tpc.nudj.ui.theme.LocalAppColors
 import com.tpc.nudj.ui.theme.NudjTheme
 
 @Composable
-fun SplashScreenLayout(
-
-){
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(LocalAppColors.current.background),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ){
-
-        Image(
-            painter = painterResource(R.drawable.splashscreenmonkey),
-            contentDescription = "Nudj monkey",
+fun SplashScreenLayout() {
+    Scaffold()
+    { paddingValues ->
+        Column(
             modifier = Modifier
-                .fillMaxWidth()
-                .weight(3f),
-            contentScale = ContentScale.Fit
+                .fillMaxSize()
+                .background(LocalAppColors.current.background),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
 
-        )
+            Image(
+                painter = painterResource(R.drawable.splash_screen_monkey),
+                contentDescription = "Nudj monkey",
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .weight(3f),
+                contentScale = ContentScale.Fit
 
-        Image(
-            painter = painterResource(R.drawable.splashscreenwelcome),
-            contentDescription = "Welcome",
-            modifier = Modifier
-                .fillMaxWidth(0.8f)
-                .height(80.dp),
-            contentScale = ContentScale.Fit
-        )
-        Spacer(modifier = Modifier.weight(0.7f))
+            )
+
+            Image(
+                painter = painterResource(R.drawable.splash_screen_welcome),
+                contentDescription = "Welcome",
+                modifier = Modifier
+                    .fillMaxWidth(0.8f)
+                    .height(80.dp),
+                contentScale = ContentScale.Fit
+            )
+            Spacer(modifier = Modifier.weight(0.7f))
+        }
     }
 }
 
@@ -59,7 +62,7 @@ fun SplashScreenLayout(
 @Composable
 private fun SplashScreenLayoutPreview() {
     NudjTheme {
-       SplashScreenLayout()
+        SplashScreenLayout()
 
     }
 }
